desafio_api_bancaria/requirements.txt
fastapi
uvicorn[standard]
databases[aiosqlite]
sqlalchemy
pyjwt
pydantic-settings
pytest
pytest-asyncio
httpx
